feat: ZettelCard 경로 변경, ZettelCard는 보여주기만 한다+onClick?
